Consolidate the stackify block walk into BlockPlanner
- UseTracker (uses.rs) owns per-block use counting, cached-immediate
  tracking, and last-use/preserve queries; BlockLiveSets is gone.
- rescue.rs owns dead-prefix cleanup and reachability-rescue
  heuristics; the ReachabilityValues wrapper is gone.
- MemState groups the driver's memory-planning borrows; MemPlan::new
  drops from nine args to four and with_planner constructs it once.
- BlockPlanner (block.rs, was block_sim.rs) owns the per-block walk
  with one method per terminator kind; observer calls live at one
  layer; edge bookkeeping stays on the driver as record_*_edge. All
  four take_stack dances are gone.
- IterationPlanner is renamed FunctionPlanner and lives in driver.rs.
