--- crates/codegen/src/stackalloc/stackify/iteration.rs
@@ -1,905 +1,0 @@
-use cranelift_entity::SecondaryMap;
-use smallvec::SmallVec;
-use sonatina_ir::{BlockId, Function, I256, InstId, ValueId};
-use std::collections::BTreeMap;
-
-use crate::{bitset::BitSet, isa::evm::immediate_materialization_code_len, stackalloc::Actions};
-
-use super::{
-    alloc::StackifyAlloc,
-    block_sim::{BlockSimState, PlannerActionSink, run_block_sim},
-    builder::{StackifyContext, StackifyReachability},
-    planner::{self, NormalizeSearchScratch, Planner},
-    slots::{FreeSlotPools, FreeSlots, SlotPool, SpillSlotPools},
-    spill::SpillSet,
-    sym_stack::{StackItem, SymStack},
-    templates::{
-        BlockTemplate, TransferOrder, canonical_transfer_order, choose_transfer, project_transfer,
-    },
-    trace::StackifyObserver,
-};
-
-pub(super) struct IterationPlanner<'a, 'ctx, O: StackifyObserver> {
-    ctx: &'a StackifyContext<'ctx>,
-    spill: SpillSet<'a>,
-    slots: &'a mut SpillSlotPools,
-    templates: &'a mut SecondaryMap<BlockId, BlockTemplate>,
-    terminal_chain_blocks: &'a BitSet<BlockId>,
-    carry_in: &'a SecondaryMap<BlockId, BitSet<ValueId>>,
-    alloc: &'a mut StackifyAlloc,
-    /// Provisional per-iteration object-id assignment (`assign_spill_obj_ids`), read by
-    /// `MemPlan` during planning before storage is finalized.
-    spill_obj: &'a SecondaryMap<ValueId, Option<crate::isa::evm::static_arena_alloc::StackObjId>>,
-    spill_requests: &'a mut BitSet<ValueId>,
-    object_spill_requests: &'a mut BitSet<ValueId>,
-    forced_object_spills: &'a BitSet<ValueId>,
-    inherited_stack: BTreeMap<BlockId, (BlockId, SymStack)>,
-    pending_edges: BTreeMap<BlockId, Vec<PendingEdge>>,
-    planned_blocks: BitSet<BlockId>,
-    search_scratch: &'a mut NormalizeSearchScratch,
-    observer: &'a mut O,
-}
-
-struct PendingEdge {
-    pred: BlockId,
-    inst: InstId,
-    stack: SymStack,
-    free_slots: FreeSlotPools,
-    action_start: usize,
-    /// Index of this edge's `DeferredExit` event in the observer's trace (0 for `NullObserver`),
-    /// used to backfill the exit fixup actions once the merge template is resolved.
-    trace_token: usize,
-}
-
-#[derive(Clone, Copy)]
-pub(super) struct ReachabilityValues<'a> {
-    pub(super) func: &'a Function,
-}
-
-impl ReachabilityValues<'_> {
-    fn retains(self, value: ValueId) -> bool {
-        !self.func.dfg.value_is_imm(value)
-    }
-}
-
-impl<'a, 'ctx, O: StackifyObserver> IterationPlanner<'a, 'ctx, O> {
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn new(
-        ctx: &'a StackifyContext<'ctx>,
-        spill: SpillSet<'a>,
-        slots: &'a mut SpillSlotPools,
-        templates: &'a mut SecondaryMap<BlockId, BlockTemplate>,
-        terminal_chain_blocks: &'a BitSet<BlockId>,
-        carry_in: &'a SecondaryMap<BlockId, BitSet<ValueId>>,
-        alloc: &'a mut StackifyAlloc,
-        spill_obj: &'a SecondaryMap<
-            ValueId,
-            Option<crate::isa::evm::static_arena_alloc::StackObjId>,
-        >,
-        spill_requests: &'a mut BitSet<ValueId>,
-        object_spill_requests: &'a mut BitSet<ValueId>,
-        forced_object_spills: &'a BitSet<ValueId>,
-        inherited_stack: BTreeMap<BlockId, (BlockId, SymStack)>,
-        search_scratch: &'a mut NormalizeSearchScratch,
-        observer: &'a mut O,
-    ) -> Self {
-        Self {
-            ctx,
-            spill,
-            slots,
-            templates,
-            terminal_chain_blocks,
-            carry_in,
-            alloc,
-            spill_obj,
-            spill_requests,
-            object_spill_requests,
-            forced_object_spills,
-            inherited_stack,
-            pending_edges: BTreeMap::new(),
-            planned_blocks: BitSet::default(),
-            search_scratch,
-            observer,
-        }
-    }
-
-    fn with_actions_planner<R>(
-        &mut self,
-        stack: &mut SymStack,
-        actions: &mut Actions,
-        free_slots: &mut FreeSlotPools,
-        f: impl FnOnce(&mut Planner) -> R,
-    ) -> R {
-        let mem = planner::MemPlan::new(
-            self.spill,
-            &mut *self.spill_requests,
-            self.ctx,
-            self.spill_obj,
-            &self.alloc.exact_local_addr,
-            self.object_spill_requests,
-            self.forced_object_spills,
-            free_slots,
-            &mut *self.slots,
-        );
-        let mut planner = Planner::new(self.ctx, stack, actions, mem, &mut *self.search_scratch);
-        f(&mut planner)
-    }
-
-    pub(super) fn plan_blocks(&mut self) {
-        for &block in self.ctx.dom.rpo() {
-            if block != self.ctx.entry && !self.ctx.dom.is_reachable(block) {
-                continue;
-            }
-
-            self.plan_block(block);
-        }
-
-        debug_assert!(
-            self.pending_edges.is_empty(),
-            "unresolved stackify edges remain"
-        );
-    }
-
-    fn plan_block(&mut self, block: BlockId) {
-        let mut free_slots: FreeSlotPools = FreeSlotPools::default();
-        let mut prologue: Actions = Actions::new();
-
-        let live_sets = BlockSimState::block_live_sets(self.ctx, block);
-        self.resolve_pending_edges(block);
-
-        let inherited = self.inherited_stack.remove(&block);
-        if self.terminal_chain_blocks.contains(block) {
-            self.freeze_template(block, TransferOrder::new());
-        } else if let Some((_pred, stack)) = inherited.as_ref() {
-            self.freeze_template_from_stack(block, stack);
-        } else if block != self.ctx.entry {
-            self.freeze_template_canonical(block);
-        }
-
-        self.observer
-            .on_block_header(self.ctx.func, block, &self.templates[block]);
-        self.planned_blocks.insert(block);
-
-        let stack = if self.terminal_chain_blocks.contains(block) {
-            SymStack::opaque_prefix_empty(self.ctx.has_internal_return)
-        } else if let Some((pred, mut inh)) = inherited {
-            // Dynamic entry stack (single predecessor).
-            if block != self.ctx.entry {
-                debug_assert_eq!(
-                    self.ctx.cfg.pred_num_of(block),
-                    1,
-                    "inherited stack implies single-predecessor block"
-                );
-                self.observer.on_block_inherited(
-                    self.ctx.func,
-                    block,
-                    pred,
-                    &inh,
-                    &live_sets.live_future,
-                    &live_sets.live_out,
-                );
-                let has_phi_params = !self.ctx.phi_results[block].is_empty();
-
-                // Single-predecessor blocks without phis do not need exact entry
-                // template normalization. Keeping the inherited stack avoids pointless bottom
-                // reshuffling that can cascade into SWAP/POP churn.
-                if has_phi_params {
-                    let tmpl = self.templates[block].clone();
-                    self.with_actions_planner(
-                        &mut inh,
-                        &mut prologue,
-                        &mut free_slots,
-                        |planner| {
-                            planner.plan_edge_fixup_to_template(&tmpl, pred, block);
-                        },
-                    );
-                }
-                self.observer.on_block_prologue(&prologue);
-            }
-            inh
-        } else {
-            SymStack::from_template(&self.templates[block], self.ctx.has_internal_return)
-        };
-
-        let state = BlockSimState::with_live_sets(block, stack, free_slots, prologue, live_sets);
-        let state = run_block_sim(self, state);
-
-        // `on_inst_start` runs for every non-phi instruction (including the terminator that every
-        // block has), so a non-empty prologue is always injected during `run_block_sim`.
-        debug_assert!(
-            state.prologue.is_empty() || state.injected_prologue,
-            "prologue was not injected during the block walk"
-        );
-    }
-
-    fn resolve_pending_edges(&mut self, block: BlockId) {
-        let Some(mut edges) = self.pending_edges.remove(&block) else {
-            return;
-        };
-        debug_assert!(
-            !self.inherited_stack.contains_key(&block),
-            "pending merge edges cannot also inherit one stack"
-        );
-        debug_assert!(
-            !self.planned_blocks.contains(block),
-            "pending edge target already planned"
-        );
-
-        let projected: Vec<(BlockId, TransferOrder)> = edges
-            .iter()
-            .map(|edge| {
-                (
-                    edge.pred,
-                    project_transfer(&edge.stack, &self.carry_in[block]),
-                )
-            })
-            .collect();
-        if !projected.is_empty() {
-            self.freeze_template(block, choose_transfer(self.ctx, block, &projected));
-        }
-
-        let tmpl = self.templates[block].clone();
-        for edge in edges.iter_mut() {
-            debug_assert_eq!(
-                self.alloc.pre_actions[edge.inst].len(),
-                edge.action_start,
-                "deferred edge action list changed before resolution"
-            );
-            self.with_planner(
-                &mut edge.stack,
-                &mut edge.free_slots,
-                PlannerActionSink::Pre(edge.inst),
-                |planner| planner.plan_edge_fixup_to_template(&tmpl, edge.pred, block),
-            );
-            self.observer.on_deferred_exit_actions(
-                edge.trace_token,
-                &self.alloc.pre_actions[edge.inst][edge.action_start..],
-            );
-        }
-    }
-
-    fn freeze_template_from_stack(&mut self, block: BlockId, stack: &SymStack) {
-        self.freeze_template(block, project_transfer(stack, &self.carry_in[block]));
-    }
-
-    fn freeze_template_canonical(&mut self, block: BlockId) {
-        let transfer = canonical_transfer_order(
-            &self.carry_in[block],
-            &self.ctx.dom_depth,
-            &self.ctx.def_info,
-        );
-        self.freeze_template(block, transfer);
-    }
-
-    fn freeze_template(&mut self, block: BlockId, transfer: TransferOrder) {
-        self.templates[block].freeze_transfer(transfer);
-    }
-
-    pub(super) fn ctx(&self) -> &StackifyContext<'ctx> {
-        self.ctx
-    }
-
-    pub(super) fn call_uses_stack_continuation(&self, inst: InstId) -> bool {
-        call_has_local_return(self.ctx.func, inst)
-    }
-
-    pub(super) fn scratch_slots(&self) -> &SlotPool {
-        &self.slots.scratch
-    }
-
-    pub(super) fn debug_assert_inst_actions_empty(&self, inst: InstId) {
-        // Within one fixed-point iteration each inst is visited once on a fresh `StackifyAlloc`,
-        // so its action buffers are always still empty when the walk reaches it.
-        debug_assert!(
-            self.alloc.pre_actions[inst].is_empty()
-                && self.alloc.post_actions[inst].is_empty()
-                && self.alloc.brtable_actions[inst].is_empty(),
-            "inst action buffers are not empty at the start of the block walk"
-        );
-    }
-
-    pub(super) fn pre_actions_len(&self, inst: InstId) -> usize {
-        self.alloc.pre_actions[inst].len()
-    }
-
-    pub(super) fn with_pre_actions<R>(
-        &mut self,
-        inst: InstId,
-        f: impl FnOnce(&mut Actions) -> R,
-    ) -> R {
-        f(&mut self.alloc.pre_actions[inst])
-    }
-
-    pub(super) fn with_planner<R>(
-        &mut self,
-        stack: &mut SymStack,
-        free_slots: &mut FreeSlotPools,
-        sink: PlannerActionSink,
-        f: impl FnOnce(&mut Planner<'_, '_>) -> R,
-    ) -> R {
-        match sink {
-            PlannerActionSink::Pre(inst) => {
-                let mem = planner::MemPlan::new(
-                    self.spill,
-                    &mut *self.spill_requests,
-                    self.ctx,
-                    self.spill_obj,
-                    &self.alloc.exact_local_addr,
-                    self.object_spill_requests,
-                    self.forced_object_spills,
-                    free_slots,
-                    &mut *self.slots,
-                );
-                let mut planner = Planner::new(
-                    self.ctx,
-                    stack,
-                    &mut self.alloc.pre_actions[inst],
-                    mem,
-                    &mut *self.search_scratch,
-                );
-                f(&mut planner)
-            }
-            PlannerActionSink::Post(inst) => {
-                let mem = planner::MemPlan::new(
-                    self.spill,
-                    &mut *self.spill_requests,
-                    self.ctx,
-                    self.spill_obj,
-                    &self.alloc.exact_local_addr,
-                    self.object_spill_requests,
-                    self.forced_object_spills,
-                    free_slots,
-                    &mut *self.slots,
-                );
-                let mut planner = Planner::new(
-                    self.ctx,
-                    stack,
-                    &mut self.alloc.post_actions[inst],
-                    mem,
-                    &mut *self.search_scratch,
-                );
-                f(&mut planner)
-            }
-            PlannerActionSink::BrTableCase { inst, case_idx } => {
-                let mut actions = Actions::new();
-                let mem = planner::MemPlan::new(
-                    self.spill,
-                    &mut *self.spill_requests,
-                    self.ctx,
-                    self.spill_obj,
-                    &self.alloc.exact_local_addr,
-                    self.object_spill_requests,
-                    self.forced_object_spills,
-                    free_slots,
-                    &mut *self.slots,
-                );
-                let result = {
-                    let mut planner = Planner::new(
-                        self.ctx,
-                        stack,
-                        &mut actions,
-                        mem,
-                        &mut *self.search_scratch,
-                    );
-                    f(&mut planner)
-                };
-                debug_assert_eq!(self.alloc.brtable_actions[inst].len(), case_idx);
-                self.alloc.brtable_actions[inst].push(actions);
-                result
-            }
-        }
-    }
-
-    pub(super) fn on_inst_start(
-        &mut self,
-        state: &mut BlockSimState,
-        inst: InstId,
-        last_use: &BitSet<ValueId>,
-    ) {
-        if !state.prologue.is_empty() && !state.injected_prologue {
-            self.alloc.pre_actions[inst].extend_from_slice(&state.prologue);
-            state.injected_prologue = true;
-        }
-        self.observer.on_inst_start(
-            self.ctx.func,
-            inst,
-            &state.stack,
-            state.live_future(),
-            state.live_out(),
-            last_use,
-        );
-    }
-
-    pub(super) fn on_cleanup_actions(&mut self, inst: InstId, start: usize, end: usize) {
-        self.observer
-            .on_inst_actions("cleanup", &self.alloc.pre_actions[inst][start..end], None);
-    }
-
-    pub(super) fn on_pre_actions(&mut self, inst: InstId, start: usize) {
-        self.observer
-            .on_inst_actions("pre", &self.alloc.pre_actions[inst][start..], None);
-    }
-
-    pub(super) fn on_post_actions(&mut self, inst: InstId) {
-        self.observer
-            .on_inst_actions("post", &self.alloc.post_actions[inst], None);
-    }
-
-    pub(super) fn on_normal_inst(&mut self, inst: InstId, args: &[ValueId], results: &[ValueId]) {
-        self.observer
-            .on_inst_normal(self.ctx.func, inst, args, results);
-    }
-
-    pub(super) fn on_return(&mut self, inst: InstId, start: usize) {
-        self.observer
-            .on_inst_actions("return", &self.alloc.pre_actions[inst][start..], None);
-        let ret_vals: SmallVec<[ValueId; 16]> = self
-            .ctx
-            .func
-            .dfg
-            .return_args(inst)
-            .map(|args| args.iter().copied().collect())
-            .unwrap_or_default();
-        self.observer
-            .on_inst_return(self.ctx.func, inst, ret_vals.as_slice());
-    }
-
-    pub(super) fn on_jump(
-        &mut self,
-        state: &mut BlockSimState,
-        inst: InstId,
-        dest: BlockId,
-        action_start: usize,
-    ) {
-        if self.terminal_chain_blocks.contains(dest) {
-        } else if self.ctx.cfg.pred_num_of(dest) > 1
-            && dest != self.ctx.entry
-            && !self.planned_blocks.contains(dest)
-        {
-            debug_assert!(
-                self.ctx.scc.is_reachable(dest),
-                "pending edge target must be reachable"
-            );
-            self.pending_edges
-                .entry(dest)
-                .or_default()
-                .push(PendingEdge {
-                    pred: state.block,
-                    inst,
-                    stack: state.stack.clone(),
-                    free_slots: state.free_slots.clone(),
-                    action_start,
-                    trace_token: self.observer.on_deferred_inst_jump(inst, dest),
-                });
-            return;
-        } else if self.ctx.cfg.pred_num_of(dest) == 1
-            && dest != self.ctx.entry
-            && !self.planned_blocks.contains(dest)
-        {
-            self.inherited_stack
-                .entry(dest)
-                .or_insert_with(|| (state.block, state.stack.clone()));
-        } else {
-            let tmpl = self.templates[dest].clone();
-            let src = state.block;
-            self.with_planner(
-                &mut state.stack,
-                &mut state.free_slots,
-                PlannerActionSink::Pre(inst),
-                |planner| planner.plan_edge_fixup_to_template(&tmpl, src, dest),
-            );
-        }
-
-        self.observer.on_inst_actions(
-            "exit",
-            &self.alloc.pre_actions[inst][action_start..],
-            Some(dest),
-        );
-        self.observer.on_inst_jump(inst, dest);
-    }
-
-    pub(super) fn on_branch(&mut self, inst: InstId, cond: ValueId, dests: &[BlockId]) {
-        self.observer.on_inst_br(self.ctx.func, inst, cond, dests);
-    }
-
-    pub(super) fn on_branch_edge(
-        &mut self,
-        state: &mut BlockSimState,
-        succ: BlockId,
-        stack: SymStack,
-    ) {
-        assert!(
-            !self.planned_blocks.contains(succ),
-            "multiway branch edge to already-planned block {succ:?}: run StackifyEdgeSplitter \
-             before stackify to split in-cycle multiway edges"
-        );
-        debug_assert_eq!(
-            self.ctx.cfg.pred_num_of(succ),
-            1,
-            "no critical edges: branch target must be single-pred"
-        );
-        self.inherited_stack
-            .entry(succ)
-            .or_insert_with(|| (state.block, stack));
-    }
-
-    pub(super) fn on_br_table_edge(
-        &mut self,
-        state: &mut BlockSimState,
-        succ: BlockId,
-        stack: SymStack,
-    ) {
-        assert!(
-            !self.planned_blocks.contains(succ),
-            "multiway br_table edge to already-planned block {succ:?}: run StackifyEdgeSplitter \
-             before stackify to split in-cycle multiway edges"
-        );
-        debug_assert_eq!(
-            self.ctx.cfg.pred_num_of(succ),
-            1,
-            "no critical edges: br_table target must be single-pred"
-        );
-        self.inherited_stack
-            .entry(succ)
-            .or_insert_with(|| (state.block, stack));
-    }
-
-    pub(super) fn on_br_table(&mut self, inst: InstId) {
-        self.observer.on_inst_br_table(inst);
-    }
-}
-
-pub(super) fn skip_pre_exit_cleanup(func: &Function, inst: InstId) -> bool {
-    func.dfg.is_exit(inst) && !func.dfg.is_return(inst)
-}
-
-pub(super) fn count_block_uses(
-    ctx: &StackifyContext<'_>,
-    block: BlockId,
-) -> (BTreeMap<ValueId, u32>, BitSet<ValueId>, BTreeMap<I256, u32>) {
-    let mut counts: BTreeMap<ValueId, u32> = BTreeMap::new();
-    let mut cached_counts: BTreeMap<I256, u32> = BTreeMap::new();
-    for inst in ctx.func.layout.iter_inst(block) {
-        if ctx.func.dfg.is_phi(inst) {
-            continue;
-        }
-        for v in operand_order_for_stackify(ctx.func, inst, &ctx.value_aliases) {
-            if ctx.retains_value(v) {
-                *counts.entry(v).or_insert(0) += 1;
-            } else if ctx.stack_caches_immediate(v)
-                && let Some(imm) = ctx.func.dfg.value_imm(v)
-            {
-                *cached_counts.entry(imm.as_i256()).or_insert(0) += 1;
-            }
-        }
-    }
-    let live_future: BitSet<ValueId> = counts.keys().copied().collect();
-    (counts, live_future, cached_counts)
-}
-
-pub(super) fn cached_immediate_preserve_values_in_inst(
-    ctx: &StackifyContext<'_>,
-    args: &[ValueId],
-    cached_remaining_uses: &BTreeMap<I256, u32>,
-) -> BitSet<ValueId> {
-    let mut inst_counts: BTreeMap<I256, u32> = BTreeMap::new();
-    for &value in args {
-        if ctx.stack_caches_immediate(value)
-            && let Some(imm) = ctx.func.dfg.value_imm(value)
-        {
-            *inst_counts.entry(imm.as_i256()).or_insert(0) += 1;
-        }
-    }
-
-    let mut preserve: BitSet<ValueId> = BitSet::default();
-    for &value in args {
-        if ctx.stack_caches_immediate(value)
-            && let Some(imm) = ctx.func.dfg.value_imm(value)
-            && cached_remaining_uses
-                .get(&imm.as_i256())
-                .copied()
-                .unwrap_or(0)
-                > inst_counts.get(&imm.as_i256()).copied().unwrap_or(0)
-        {
-            preserve.insert(value);
-        }
-    }
-    preserve
-}
-
-pub(super) fn consume_cached_immediate_uses(
-    ctx: &StackifyContext<'_>,
-    args: &[ValueId],
-    cached_remaining_uses: &mut BTreeMap<I256, u32>,
-) {
-    for &value in args {
-        if ctx.stack_caches_immediate(value)
-            && let Some(imm) = ctx.func.dfg.value_imm(value)
-            && let Some(count) = cached_remaining_uses.get_mut(&imm.as_i256())
-        {
-            *count = count.saturating_sub(1);
-        }
-    }
-    cached_remaining_uses.retain(|_, count| *count != 0);
-}
-
-fn pop_dead_tops(
-    stack: &mut SymStack,
-    live_future: &BitSet<ValueId>,
-    live_out: &BitSet<ValueId>,
-    actions: &mut Actions,
-) {
-    while let Some(StackItem::Value(v)) = stack.top() {
-        if live_future.contains(*v) || live_out.contains(*v) {
-            break;
-        }
-        stack.pop(actions);
-    }
-}
-
-pub(super) fn clean_dead_stack_prefix(
-    reach: StackifyReachability,
-    stack: &mut SymStack,
-    live_future: &BitSet<ValueId>,
-    live_out: &BitSet<ValueId>,
-    actions: &mut Actions,
-) {
-    // Two local cleanups:
-    // - pop any dead values that reach the top
-    // - if a live value is on top and there is a contiguous dead chain directly beneath it,
-    //   swap the live value down and pop the dead chain off the top.
-    loop {
-        let before_len = stack.len();
-        pop_dead_tops(stack, live_future, live_out, actions);
-
-        // If the top is not a normal value, don't try to reorder under it.
-        let Some(StackItem::Value(top)) = stack.top() else {
-            break;
-        };
-        // `pop_dead_tops` only stops on a live value (or a non-`Value`, handled above).
-        debug_assert!(
-            live_future.contains(*top) || live_out.contains(*top),
-            "pop_dead_tops left a dead value on top"
-        );
-
-        let is_dead = |v: ValueId| !live_future.contains(v) && !live_out.contains(v);
-        let dead_run = stack
-            .iter()
-            .skip(1)
-            .take_while(|&v| matches!(v, StackItem::Value(v) if is_dead(*v)))
-            .count();
-        if dead_run == 0 {
-            break;
-        }
-
-        // Swap the top live value with the deepest dead value in the contiguous chain (within
-        // SWAP16 reach), then pop that chunk off. Repeat until the chain is gone.
-        let mut remaining = dead_run;
-        while remaining > 0 {
-            let swap_depth = remaining.min(reach.swap_max.saturating_sub(1));
-            stack.swap(swap_depth, actions);
-            stack.pop_n(swap_depth, actions);
-            remaining -= swap_depth;
-        }
-
-        if stack.len() == before_len {
-            break;
-        }
-    }
-}
-
-pub(super) fn operand_order_for_stackify(
-    func: &Function,
-    inst: InstId,
-    value_aliases: &SecondaryMap<ValueId, Option<ValueId>>,
-) -> SmallVec<[ValueId; 8]> {
-    let mut args: SmallVec<[ValueId; 8]> = func
-        .dfg
-        .inst(inst)
-        .collect_values()
-        .into_iter()
-        .map(|v| value_aliases[v].unwrap_or(v))
-        .collect();
-
-    if call_has_local_return(func, inst) && !args.is_empty() {
-        // Rotation pairs with `Planner::prepare_internal_call`; see the ABI note there.
-        args.as_mut_slice().rotate_left(1);
-    }
-
-    args
-}
-
-fn call_has_local_return(func: &Function, inst: InstId) -> bool {
-    func.dfg.call_info(inst).is_some_and(|call| {
-        func.ctx()
-            .func_effects(call.callee())
-            .may_return_to_caller()
-    })
-}
-
-pub(super) fn consume_operand_uses(
-    ctx: &StackifyContext<'_>,
-    args: &[ValueId],
-    remaining_uses: &mut BTreeMap<ValueId, u32>,
-    live_future: &mut BitSet<ValueId>,
-    live_out: &BitSet<ValueId>,
-    scratch_slots: &SlotPool,
-    free_scratch_slots: &mut FreeSlots,
-) {
-    for &v in args {
-        if ctx.retains_value(v)
-            && let Some(n) = remaining_uses.get_mut(&v)
-        {
-            let before = *n;
-            *n = n.saturating_sub(1);
-            if before != 0 && *n == 0 {
-                live_future.remove(v);
-                if !ctx.func.dfg.value_is_imm(v) && !live_out.contains(v) {
-                    scratch_slots.release_if_assigned(v, free_scratch_slots);
-                }
-            }
-        }
-    }
-}
-
-pub(super) fn last_use_values_in_inst(
-    ctx: &StackifyContext<'_>,
-    args: &[ValueId],
-    remaining_uses: &BTreeMap<ValueId, u32>,
-    live_out: &BitSet<ValueId>,
-) -> BitSet<ValueId> {
-    let mut inst_counts: BTreeMap<ValueId, u32> = BTreeMap::new();
-    for &v in args.iter() {
-        if !ctx.retains_value(v) {
-            continue;
-        }
-        *inst_counts.entry(v).or_insert(0) += 1;
-    }
-
-    let mut last_use: BitSet<ValueId> = BitSet::default();
-    for (v, count) in inst_counts {
-        let rem = remaining_uses.get(&v).copied().unwrap_or(0);
-        if rem == count && !live_out.contains(v) {
-            last_use.insert(v);
-        }
-    }
-    last_use
-}
-
-pub(super) fn improve_reachability_before_operands(
-    values: ReachabilityValues<'_>,
-    args: &[ValueId],
-    reach: StackifyReachability,
-    stack: &mut SymStack,
-    live_future: &BitSet<ValueId>,
-    live_out: &BitSet<ValueId>,
-    actions: &mut Actions,
-) {
-    const AGGRESSIVE_REACHABILITY_DEPTH: usize = 20;
-
-    let mut protected_args: BitSet<ValueId> = BitSet::default();
-    for &arg in args.iter() {
-        if values.retains(arg) {
-            protected_args.insert(arg);
-        }
-    }
-
-    // If at least one operand is on the stack but unreachable by `DUP16`, attempt a more
-    // aggressive cleanup to bring it back into reach. This extends slightly past `SWAP16` reach
-    // by allowing deletions that shift the stack (e.g. popping dead/cheap values above an
-    // operand at depth 18-20).
-    let mut needs_aggressive = false;
-    for &arg in args.iter() {
-        if values.retains(arg)
-            && stack.find_reachable_value(arg, reach.dup_max).is_none()
-            && stack
-                .find_reachable_value(arg, AGGRESSIVE_REACHABILITY_DEPTH)
-                .is_some()
-        {
-            needs_aggressive = true;
-            break;
-        }
-    }
-    if !needs_aggressive {
-        return;
-    }
-
-    // Bound the amount of cleanup we do per instruction to avoid pathological swap chains.
-    const MAX_DELETIONS: usize = 8;
-    let mut deletions: usize = 0;
-
-    while deletions < MAX_DELETIONS {
-        let mut progressed = false;
-
-        for &arg in args.iter() {
-            if values.retains(arg)
-                && stack.find_reachable_value(arg, reach.dup_max).is_none()
-                && let Some(pos) = stack.find_reachable_value(arg, AGGRESSIVE_REACHABILITY_DEPTH)
-                && let Some(victim) = choose_reachability_victim(
-                    values,
-                    stack,
-                    pos,
-                    &protected_args,
-                    reach,
-                    live_future,
-                    live_out,
-                )
-            {
-                stack.stable_delete_at_depth(victim + 1, actions);
-                deletions += 1;
-                progressed = true;
-                break;
-            }
-        }
-
-        if !progressed {
-            break;
-        }
-    }
-}
-
-fn choose_reachability_victim(
-    values: ReachabilityValues<'_>,
-    stack: &SymStack,
-    above: usize,
-    protected_args: &BitSet<ValueId>,
-    reach: StackifyReachability,
-    live_future: &BitSet<ValueId>,
-    live_out: &BitSet<ValueId>,
-) -> Option<usize> {
-    let limit = stack.len_above_func_ret().min(reach.swap_max);
-    let above = above.min(limit);
-
-    // 1) Prefer deleting dead values, starting from the shallowest depth to minimize `SWAP*`
-    // chains. This includes immediates: if they're dead, removing them cannot introduce new
-    // rematerialization cost.
-    for (i, item) in stack.iter().take(above).enumerate() {
-        if let StackItem::Value(v) = item
-            && !protected_args.contains(*v)
-            && !live_future.contains(*v)
-            && !live_out.contains(*v)
-        {
-            return Some(i);
-        }
-    }
-
-    // 2) Then evict "cheap" immediates (they are always rematerializable).
-    for (i, item) in stack.iter().take(above).enumerate() {
-        if let StackItem::Value(v) = item
-            && !protected_args.contains(*v)
-            && values.func.dfg.value_is_imm(*v)
-            && is_evictable_imm(values.func, *v)
-        {
-            return Some(i);
-        }
-    }
-
-    // 3) Then delete redundant duplicates of non-operands (keeping the shallowest copy).
-    let mut first_index: BTreeMap<ValueId, usize> = BTreeMap::new();
-    for (i, item) in stack.iter().take(above).enumerate() {
-        let StackItem::Value(v) = item else {
-            continue;
-        };
-        first_index.entry(*v).or_insert(i);
-    }
-
-    for (i, item) in stack.iter().take(above).enumerate() {
-        if let StackItem::Value(v) = item
-            && !protected_args.contains(*v)
-            && let Some(&first) = first_index.get(v)
-            && first != i
-        {
-            return Some(i);
-        }
-    }
-
-    None
-}
-
-fn is_evictable_imm(func: &Function, v: ValueId) -> bool {
-    const MAX_MATERIALIZATION_BYTES: usize = 3; // PUSH2 or smaller.
-    let Some(imm) = func.dfg.value_imm(v) else {
-        return false;
-    };
-    immediate_materialization_code_len(imm) <= MAX_MATERIALIZATION_BYTES
-}
